changed double to float

+++ src/FFNN/NeuralNetwork.java
@@ -27,10 +27,10 @@ public class NeuralNetwork {
             }
 
             // Force the bias node's output value to 1.0. It's last neuron created above
-            m_layers.get(m_layers.size()-1).peekLast().setOutputValue(1.0);
+            m_layers.get(m_layers.size()-1).peekLast().setOutputValue(1.0f);
         }
     }
-    public void feedForward(ArrayList<Double> inputValues)
+    public void feedForward(ArrayList<Float> inputValues)
     {
         assert(inputValues.size() == m_layers.get(0).size() - 1);
 
@@ -50,25 +50,25 @@ public class NeuralNetwork {
             }
         }
     }
-    public void backProp(ArrayList<Double> targetValues)
+    public void backProp(ArrayList<Float> targetValues)
     {
         // Calculate overall net error (RMS of output neuron errors)
         Layer outputLayer = m_layers.get(m_layers.size()-1);
-        m_error = 0.0;
+        m_error = 0.0f;
 
         for (int n = 0; n < outputLayer.size() - 1; n++)
         {
-            double delta = targetValues.get(n) - outputLayer.get(n).getOutputValue();
+            float delta = targetValues.get(n) - outputLayer.get(n).getOutputValue();
             m_error += delta * delta;
         }
         m_error /= outputLayer.size() - 1; //get average errorsquared
-        m_error = Math.sqrt(m_error); // RMS
+        m_error = (float)Math.sqrt(m_error); // RMS
 
         // Implement a recent average measurement;
 
         m_recentAverageError =
                 (m_recentAverageError * m_recentAverageSmoothingFactor + m_error)
-                        / (m_recentAverageSmoothingFactor + 1.0);
+                        / (m_recentAverageSmoothingFactor + 1.0f);
 
         // Calculate output layer gradients
         for (int n = 0; n < outputLayer.size() - 1; n++)
@@ -102,7 +102,7 @@ public class NeuralNetwork {
             }
         }
     }
-    public void getResults(ArrayList<Double> resultValues)
+    public void getResults(ArrayList<Float> resultValues)
     {
         resultValues.clear();
 
@@ -111,11 +111,11 @@ public class NeuralNetwork {
             resultValues.add(m_layers.get(m_layers.size()-1).get(n).getOutputValue());
         }
     }
-    public double getNeuronOutput(int x, int y)
+    public float getNeuronOutput(int x, int y)
     {
         return m_layers.get(x).get(y).getOutputValue();
     }
-    public double getRecentAverageError() { return m_recentAverageError; }
+    public float getRecentAverageError() { return m_recentAverageError; }
 
     public void saveNeuronWeights()
     {
@@ -146,8 +146,8 @@ public class NeuralNetwork {
     }
 
     private ArrayList<Layer> m_layers; // m_layers[layerNum][neuronNum]
-    private double m_error;
-    private double m_recentAverageError;
-    private double m_recentAverageSmoothingFactor;
+    private float m_error;
+    private float m_recentAverageError;
+    private float m_recentAverageSmoothingFactor;
 
 }
